Parsing
Goes over how to parse by using an example of an input text file
providing row and col dimensions. The input text file has a grid which
is then converted to a String[][]. This grid is then converted to an
int[][], before finally being written to an output file, but not before
having each int being converted to a String.